% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpai-package.R
\docType{package}
\name{helpai-package}
\alias{helpai}
\alias{helpai-package}
\title{'SciViews::R' - Simplified Help Pages in R Thanks to AI}
\description{
This package builds custom R help pages, based on questions you ask to a LLM
(by default, it is Ollama running locally).
}
\section{Important functions}{

\itemize{
\item \code{\link[=ai_ask]{ai_ask()}} the main function for queries,
\item \code{\link[=ai_explain_term]{ai_explain_term()}} to get a definition and brief explanation on a
(statistical) term.
\item \code{\link[=ai_explain_function]{ai_explain_function()}} produces a short help page for an R
function with explanation of its main arguments and a short example.
\item \code{\link[=ai_explain_code]{ai_explain_code()}} details how a small piece of R code works.
\item \code{\link[=ai_explain_error]{ai_explain_error()}} explain what en R error message means.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/helpai}
  \item \url{https://www.sciviews.org/helpai/}
  \item \url{https://sciviews.r-universe.dev/helpai}
  \item Report bugs at \url{https://github.com/SciViews/helpai/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
